+++ docs/en/UI/Angular/Abp-Window-Service.md
@@ -1,5 +1,8 @@
-# File Utils Service
-FileUtilsService is an Angular service designed to provide utility methods related to file operations. The service has a `downloadBlob` function, which is used for downloading blobs as files within the context of a web application.
+# Abp Window Service
+
+
+## Download Blob as File 
+AbpWindowService is an Angular service designed to provide utility methods related to window operations. The service has a `downloadBlob` function, which is used for downloading blobs as files within the context of a web application.
 
 ## Usage
 
@@ -9,11 +12,11 @@ Injection:
 Firstly, ensure that the service is injected into the component or any other Angular entity where you wish to use it.
 
 ```js
-import { FileUtilsService } from '@abp/ng.core';
+import { AbpWindowService } from '@abp/ng.core';
 
-constructor(private fileUtils: FileUtilsService) { }
+constructor(private abpWindowService: AbpWindowService) { }
 // or
-// private fileUtils   = inject(FileUtilsService)
+// private abpWindowService   = inject(AbpWindowService)
 ```
 
 Downloading a Blob:
@@ -23,7 +26,7 @@ Once you have the service injected, you can use the downloadBlob method to initi
 ```js
 someMethod() {
   const myBlob = new Blob(["Hello, World!"], { type: "text/plain" });
-  this.fileUtils.downloadBlob(myBlob, "hello.txt");
+  this.abpWindowService.downloadBlob(myBlob, "hello.txt");
 }
 ```
