+++ npm/ng-packs/packages/core/src/lib/services/window.service.ts
@@ -3,7 +3,8 @@ import { DOCUMENT } from '@angular/common';
 
 @Injectable({ providedIn: 'root' })
 export class AbpWindowService {
-  protected readonly window = inject(DOCUMENT).defaultView;
+  protected readonly document = inject(DOCUMENT);
+  protected readonly window = this.document.defaultView;
   protected readonly navigator = this.window.navigator;
 
   copyToClipboard(text: string): Promise<void> {
@@ -16,5 +17,22 @@ export class AbpWindowService {
 
   reloadPage(): void {
     this.window.location.reload();
+  }
+  downloadBlob(blob: Blob, fileName: string) {
+    const blobUrl = this.window.URL.createObjectURL(blob);
+    const a = this.document.createElement('a');
+    a.style.display = 'none';
+    a.href = blobUrl;
+    a.download = fileName;
+    this.document.body.appendChild(a);
+    a.dispatchEvent(
+      new MouseEvent('click', {
+        bubbles: true,
+        cancelable: true,
+        view: this.window,
+      }),
+    );
+    this.window.URL.revokeObjectURL(blobUrl);
+    this.document.body.removeChild(a);
   }
 }
